# Make test 02473_optimize_old_parts less flaky