--- a/tests/queries/0_stateless/02473_optimize_old_parts.reference
+++ b/tests/queries/0_stateless/02473_optimize_old_parts.reference
@@ -8,5 +8,3 @@ With merge replicated any part range
 1
 With merge replicated partition only
 1
-With merge partition only and new parts
-3