… tests/queries/0_stateless/02473_optimize_old_parts.sql