### Changelog category (leave one):
- Not for changelog (changelog entry is not required)

A failure with MSAN: https://s3.amazonaws.com/clickhouse-test-reports/47290/68dea487b093c46de4d35a8aef783e4740abfd22/stateless_tests__msan__[5/6].html

A failure with S3 storage: https://s3.amazonaws.com/clickhouse-test-reports/0/581f3af9211c82705184dd1e3b7831299dc262cf/stateless_tests__release__s3_storage_.html

I couldn't reproduce it locally (*), so didn't directly check whether this change improves anything.

The test sleeps for 9 seconds and expects a background merge to happen during that time. The time before merge starts is merge_selecting_sleep_ms (5s) + min_age_to_force_merge_seconds (3s) = 8 seconds in the worst case, AFAICT. So it leaves only 1 second of margin.

This PR reduces the two settings to 1s + 1s, and the sleep to 8s, so a 6s margin. Hope that's enough.

Is it enough in S3 though? How consistent is the latency there? Does it break a lot of tests like this one?

Would be better if the test actually waited for the merges instead of the sleep & pray. Is there a good way to do that? I'm a noob.

(*) Didn't try to repro with S3. Tried running 100 instances in parallel x 100 attempts, with 50 while(true) loops on the side 100%ing the cpu - still nothing :(