@al13n321 2023-03-10
Alrighty, here's a bash script, seems better.

---

I don't understand the last test:
```
SELECT 'With merge partition only and new parts';

CREATE TABLE test_with_merge (i Int64) ENGINE = MergeTree ORDER BY i
SETTINGS min_age_to_force_merge_seconds=3, min_age_to_force_merge_on_partition_only=true;
SYSTEM STOP MERGES test_with_merge;
-- These three parts will have min_age=6 at the time of merge
INSERT INTO test_with_merge SELECT 1;
INSERT INTO test_with_merge SELECT 2;
SELECT sleepEachRow(1) FROM numbers(9) FORMAT Null;
-- These three parts will have min_age=0 at the time of merge
-- and so, nothing will be merged.
INSERT INTO test_with_merge SELECT 3;
SYSTEM START MERGES test_with_merge;

SELECT count(*) FROM system.parts WHERE database = currentDatabase() AND table='test_with_merge' AND active;
```

Comments say "three parts", but there are 2 or 1. The query at the end happens without any sleep, so merges don't get a chance to happen. I guess the idea is to check that the first 2 parts get merged, but the 3rd doesn't, because it's too new? If so, idk how to do that without relying on short sleeps.

Removed that test for now, lmk if you have better ideas.