@antonio2368 2023-03-13
> Comments say "three parts", but there are 2 or 1. The query at the end happens without any sleep, so merges don't get a chance to happen. I guess the idea is to check that the first 2 parts get merged, but the 3rd doesn't, because it's too new? If so, idk how to do that without relying on short sleeps.

the comment is wrong, the first should `two parts`

```
SYSTEM STOP MERGES test_with_merge; <--- we stop merges
-- These three parts will have min_age=6 at the time of merge. <--- should be two parts
INSERT INTO test_with_merge SELECT 1;                     
INSERT INTO test_with_merge SELECT 2; <-- we do 2 inserts to generate 2 parts
SELECT sleepEachRow(1) FROM numbers(9) FORMAT Null; <-- sleep to make them older but merges are stopped
-- These three parts will have min_age=0 at the time of merge
-- and so, nothing will be merged.
INSERT INTO test_with_merge SELECT 3; <-- insert third part
SYSTEM START MERGES test_with_merge; <-- start merges but verify we didn't merge the entire partition because the third part is too new (the goal of the test)
```